Add more tags from the spec
There are some more tags from the specification. I'm not a Go-expert. If the room_name is undefined, will the Room_Name in the struct undefined too or will this cause an error? In the spec: https://matrix.org/docs/spec/push_gateway/r0.1.0.html all of the parameters (not devices) are optional.